ci: Add release workflow for draft release with all sources as asset
Create a GitHub workflow that runs when new tags are created, in order
to create a draft release containing a new asset: an archive with all
sources required to build veristat, in other words, the sources of
veristat and of its libbpf submodule.

The draft release is at https://github.com/libbpf/veristat/releases, and
needs to be published manually. Note that the release author will be
"github-actions" instead of the person who manually publishes from the
draft.

The motivation for adding an archive with the sources of both veristat
and libbpf is to help with packaging: there is only one archive to
handle to get all the sources, without having to care about git
submodules.

Signed-off-by: Quentin Monnet <quentin@isovalent.com>